refactor(dataservice): move shared timezone helpers to time_utils
Agent-Logs-Url: https://github.com/Mites-io/BuildingDepot-v3/sessions/c33178a1-5ab9-441f-a1d1-79686c5f24b2
